refactor: organize apis, models, schemas in different dictories
