+++ backend/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "backend"
+version = "0.1.0"
+description = "Add your description here"
+requires-python = ">=3.12"
+dependencies = [
+    "alembic>=1.16.4",
+    "fastapi>=0.116.1",
+    "google-generativeai>=0.8.5",
+    "httpx>=0.28.1",
+    "psycopg2-binary>=2.9.10",
+    "pydantic>=2.11.7",
+    "pytest>=8.4.1",
+    "python-dotenv>=1.1.1",
+    "python-multipart>=0.0.20",
+    "sqlalchemy>=2.0.41",
+    "uvicorn[standard]>=0.35.0",
+]
