--- backend/requirements.txt
@@ -1,11 +1,0 @@
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-pydantic==2.5.0
-google-generativeai==0.3.2
-python-multipart==0.0.6
-python-dotenv==1.0.0
-psycopg2-binary==2.9.9
-sqlalchemy==2.0.23
-alembic==1.12.1
-pytest==7.4.3
-httpx==0.25.2
