+++ backend/src/database.py
@@ -1,9 +1,8 @@
-from sqlalchemy import create_engine, Column, String, Text, DateTime, JSON
-from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from datetime import datetime
 import os
 from dotenv import load_dotenv
+from .schemas.digest import Base
 
 load_dotenv()
 
@@ -16,18 +15,6 @@ engine = create_engine(
 )
 
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-Base = declarative_base()
-
-class Digest(Base):
-    __tablename__ = "digests"
-    
-    id = Column(String, primary_key=True, index=True)
-    public_id = Column(String, unique=True, index=True)
-    original_transcript = Column(Text, nullable=False)
-    overview = Column(Text, nullable=False)
-    key_decisions = Column(JSON, nullable=False)
-    action_items = Column(JSON, nullable=False)
-    created_at = Column(DateTime, default=datetime.utcnow)
 
 def create_tables():
     Base.metadata.create_all(bind=engine)
