+++ backend/src/models/__init__.py
@@ -1,0 +1,3 @@
+from .digest import TranscriptRequest, DigestResponse
+
+__all__ = ["TranscriptRequest", "DigestResponse"]
