+++ backend/src/schemas/__init__.py
@@ -1,0 +1,3 @@
+from .digest import Digest
+
+__all__ = ["Digest"]
